requirements.txt
PyPDF2
pdfminer.six
pdfplumber
spacy